Adding more repository tests
